refactor: attributesをauthProviderで共通化
