+++ client/src/components/pages/EditArticlePage/index.tsx
@@ -27,15 +27,10 @@ const contentOnNotEditable = (
 const EditArticlePage = () => {
   const { articleId } = useParams();
   const { contract } = useWeb3Context();
-  const { attributes } = useAuthContext();
+  const {
+    attributes: { walletAddress },
+  } = useAuthContext();
   const [ownerIdOfArticle, setOwnerIdOfArticle] = useState<null | string>(null);
-
-  const walletAddress = useMemo(() => {
-    const walletAddress = attributes.find(
-      (atr) => atr.Name === "custom:wallet_address"
-    )?.Value;
-    return walletAddress || "";
-  }, [attributes]);
 
   const isAuthorized = useMemo(() => {
     return ownerIdOfArticle === walletAddress;
